+++ test/clj_kondo/core_test.clj
@@ -246,18 +246,19 @@
                 (fn [entry-map]
                   (swap! calls conj entry-map))
                 {})]
-      (is (= 7 (:files (:summary res))))
+      (is (= 8 (:files (:summary res))))
       (assert-submaps
-        #{{:filename "corpus/use.clj" :uri #"file:/.*/corpus/use.clj" :total-files 7}
-          {:filename "corpus/case.clj" :uri #"file:/.*/corpus/case.clj" :total-files 7}
-          {:filename "corpus/schema/defs.clj" :uri #"file:/.*/corpus/schema/defs.clj" :total-files 7}
-          {:filename "corpus/schema/defmethod.clj" :uri #"file:/.*/corpus/schema/defmethod.clj" :total-files 7}
-          {:filename "corpus/schema/calls.clj" :uri #"file:/.*/corpus/schema/calls.clj" :total-files 7}
-          {:filename "corpus/schema/defrecord.clj" :uri #"file:/.*/corpus/schema/defrecord.clj" :total-files 7}
-          {:filename "dirinjar.clj/arity.clj" :uri #"jar:file:/.*/corpus/withcljdir.jar!/dirinjar.clj/arity.clj" :total-files 7}}
+        #{{:filename "corpus/use.clj" :uri #"file:/.*/corpus/use.clj" :total-files 8}
+          {:filename "corpus/case.clj" :uri #"file:/.*/corpus/case.clj" :total-files 8}
+          {:filename "corpus/schema/defs.clj" :uri #"file:/.*/corpus/schema/defs.clj" :total-files 8}
+          {:filename "corpus/schema/defmethod.clj" :uri #"file:/.*/corpus/schema/defmethod.clj" :total-files 8}
+          {:filename "corpus/schema/calls.clj" :uri #"file:/.*/corpus/schema/calls.clj" :total-files 8}
+          {:filename "corpus/schema/defrecord.clj" :uri #"file:/.*/corpus/schema/defrecord.clj" :total-files 8}
+          {:filename "corpus/schema/defprotocol.clj" :uri #"file:/.*/corpus/schema/defprotocol.clj" :total-files 8}
+          {:filename "dirinjar.clj/arity.clj" :uri #"jar:file:/.*/corpus/withcljdir.jar!/dirinjar.clj/arity.clj" :total-files 8}}
         (set @calls))
       (is (every? #(and (int? (:total-files %))
-                        (<= (:total-files %) 7)) @calls))))
+                        (<= (:total-files %) 8)) @calls))))
 
   (testing "when lint is classpath"
     (let [calls (atom [])
@@ -291,18 +292,19 @@
                 (fn [entry-map]
                   (swap! calls conj entry-map))
                 {:parallel true})]
-      (is (= 7 (:files (:summary res))))
+      (is (= 8 (:files (:summary res))))
       (assert-submaps
-        #{{:filename "corpus/use.clj" :uri #"file:/.*/corpus/use.clj" :total-files 7}
-          {:filename "corpus/schema/defs.clj" :uri #"file:/.*/corpus/schema/defs.clj" :total-files 7}
-          {:filename "corpus/case.clj" :uri #"file:/.*/corpus/case.clj" :total-files 7}
-          {:filename "corpus/schema/defmethod.clj" :uri #"file:/.*/corpus/schema/defmethod.clj" :total-files 7}
-          {:filename "corpus/schema/calls.clj" :uri #"file:/.*/corpus/schema/calls.clj" :total-files 7}
-          {:filename "corpus/schema/defrecord.clj" :uri #"file:/.*/corpus/schema/defrecord.clj" :total-files 7}
-          {:filename "dirinjar.clj/arity.clj" :uri #"jar:file:/.*/corpus/withcljdir.jar!/dirinjar.clj/arity.clj" :total-files 7}}
+        #{{:filename "corpus/use.clj" :uri #"file:/.*/corpus/use.clj" :total-files 8}
+          {:filename "corpus/schema/defs.clj" :uri #"file:/.*/corpus/schema/defs.clj" :total-files 8}
+          {:filename "corpus/case.clj" :uri #"file:/.*/corpus/case.clj" :total-files 8}
+          {:filename "corpus/schema/defmethod.clj" :uri #"file:/.*/corpus/schema/defmethod.clj" :total-files 8}
+          {:filename "corpus/schema/calls.clj" :uri #"file:/.*/corpus/schema/calls.clj" :total-files 8}
+          {:filename "corpus/schema/defrecord.clj" :uri #"file:/.*/corpus/schema/defrecord.clj" :total-files 8}
+          {:filename "corpus/schema/defprotocol.clj" :uri #"file:/.*/corpus/schema/defprotocol.clj" :total-files 8}
+          {:filename "dirinjar.clj/arity.clj" :uri #"jar:file:/.*/corpus/withcljdir.jar!/dirinjar.clj/arity.clj" :total-files 8}}
         (set @calls))
       (is (every? #(and (int? (:total-files %))
-                        (<= (:total-files %) 7)) @calls)))))
+                        (<= (:total-files %) 8)) @calls)))))
 
 ;;;; Scratch